feat: reorganize component structure and remove obsolete images
- Updated the import paths in App.jsx to reflect the new folder structure.
- Removed unused images of muscle groups and routines.
- Added new components such as Login, BlogCard, BlogPost, AuthNavbar, Background, Navbar, and pages for Blog, Exercises, and Routines.
